Separate input_filter.go from filter.go
Separate input_filter.go for input events.
